+++ src/esthri/src/ops/mod.rs
@@ -1,1 +1,2 @@
 pub(super) mod download;
+pub(super) mod upload;
